chainstats: refactor
Chains.goto now features calculate_chain_stats (`default=False`) to
calculate chain statistics on demand.

Chain uniqueness is always calculated and appended to a growing array.
This remove a lot of overhead and makes plotting of in-depth statistics
optional.

Plotting functions are cleaned but not finished. Added `report_plot`
class attribute to disable plots in the default report.